refactor: move all transaction db protocols to specific folders
