+++ src/data/protocols/db/transaction/create-transaction-repository.ts
@@ -1,4 +1,4 @@
-import { TransactionModel } from "../../../domain/models/transaction";
+import { TransactionModel } from "../../../../domain/models/transaction";
 
 export interface CreateTransactionRepository {
   create(transaction: TransactionModel): Promise<TransactionModel>;
