+++ src/data/protocols/db/transaction/find-transaction-by-filters-repository.ts
@@ -1,4 +1,4 @@
-import { TransactionModel } from "../../../domain/models/transaction";
+import { TransactionModel } from "../../../../domain/models/transaction";
 import { TransactionFiltersModel } from "src/domain/usecases/transaction/find-transaction-by-filters";
 
 export interface FindTransactionByFiltersRepository {
