+++ src/data/protocols/db/transaction/find-transaction-by-id-repository.ts
@@ -1,4 +1,4 @@
-import { TransactionModel } from "../../../domain/models/transaction";
+import { TransactionModel } from "../../../../domain/models/transaction";
 
 export interface FindTransactionByIdRepository {
   findById(_id: string): Promise<TransactionModel>;
