+++ src/data/protocols/db/transaction/get-total-amount-repository.ts
@@ -1,4 +1,4 @@
-import { TotalAmountModel } from "../../../domain/models/total-amount";
+import { TotalAmountModel } from "../../../../domain/models/total-amount";
 
 export interface GetTotalAmountRepository {
   get(): Promise<TotalAmountModel>;
