+++ src/data/usecases/db-delete-transaction.ts
@@ -1,5 +1,5 @@
 import { DeleteTransaction } from "../../domain/usecases/transaction/delete-transaction";
-import { DeleteTransactionRepository } from "../protocols/db/delete-transaction-repository";
+import { DeleteTransactionRepository } from "../protocols/db/transaction/delete-transaction-repository";
 
 export class DbDeleteTransaction implements DeleteTransaction {
   constructor(private readonly repository: DeleteTransactionRepository) {}
