+++ src/data/usecases/db-find-transaction-by-id.ts
@@ -1,6 +1,6 @@
 import { TransactionModel } from "src/domain/models/transaction";
 import { FindTransactionById } from "../../domain/usecases/transaction/find-transaction-by-id";
-import { FindTransactionByIdRepository } from "../protocols/db/find-transaction-by-id-repository";
+import { FindTransactionByIdRepository } from "../protocols/db/transaction/find-transaction-by-id-repository";
 
 export class DbFindTransactionById implements FindTransactionById {
   constructor(private readonly repository: FindTransactionByIdRepository) {}
