+++ tests/data/usecases/db-find-transaction-by-filters.spec.ts
@@ -1,5 +1,5 @@
 import { DbFindTransactionByFilters } from "../../../src/data/usecases/db-find-transaction-by-filters";
-import { FindTransactionByFiltersRepository } from "../../../src/data/protocols/db/find-transaction-by-filters-repository";
+import { FindTransactionByFiltersRepository } from "../../../src/data/protocols/db/transaction/find-transaction-by-filters-repository";
 import { TransactionModel } from "../../../src/domain/models/transaction";
 
 const makeTransactionFilter = (): {
